+++ CHANGELOG.md
@@ -33,6 +33,7 @@ whitespace characters.
 invalid address or is stopped due to any other error.
 - Fix a bug where environments could not be created with sensuctl create
 - StatsD listener on Windows is functional
+- Add version output for dev and nightly builds (#1320).
 
 ## [2.0.0-beta.1] - 2018-05-07
 ### Added